Reworked ConsoleReporter and run for Jasmine 1.3.1
I did this because we found that the current master of Jasmine is somewhat
broken in terms of functionality and definitely is lacking documentation in
terms of actually using it. Therefore, I switched to the latest stable release
v1.3.1 of Jasmine although there was a significant amount different. This
version has a completely different way of run the specs, it has no concept of
pending tests, it has a completely different API for reporting. Therefore, I
had to rework the run.erb and the ConsoleReporter to work properly with the
concepts of Jasmine v1.3.1 and the APIs.