Added route to web app to serve files in cwd.
This should resolves #12 by adding a route that makes /:path use the
snapdragon commands current working directory as the base path and respond
with the files based on :path. This is especially useful for accessing
spec/fixtures/some_fixture.txt when your Jasmine specs are stored in
spec/javascripts/some_spec.js.