Added stack traces to console reporter.
I did this so that we would be able to get just as much info out of running
the tests on the command line as we would from running the tests in the
browser. Currently, it isn't outputing in the browser even though the code is
there because for some reason the v1.3.1 Jasmine code Error objects aren't
getting the stack propery set in PhantomJS. If we run it in Chrome and watch
the console logs we see it work properly there. The interesting thing is that
before when I was using the Jasmine master branch instead of v1.3.1 I was able
to get the stack trace to work in the same version of PhantomJS. I am guessing
that v1.3.1 has an older version of whatever code supports that. I will need
to dig into it and see if I can fix it easily.