Switched from inline to script tags in runner.
I did this because this is really the suggested usage by Jasmine and I believe
that it will give us more accurate and valuable stack traces when errors
occur and maybe even correct line numbers. This resolves #13. This also
eliminates the issue with inlining where you couldn't have a <script> tag in
your javascript comments because it would close the actual parenting script
tag early.