Made snapdragon console reporter to output stack.
I did this because it was apparently working before my test case just wasn't
valid because it was a match failure result and not an exception that was
caught. Therefore, I changed the example spec to include two failures. One
that is a match failure and one is an exception which should have a stack
trace.