require_relative directives work in required files.
I added this feature so that the model of using a spec_helper.js that has
additional requires in it for the dependencies can be defined. Then that
spec_helper.js can be required in each of the spec test files. This allows
users to control the order of the required files in the spec_helper.js. This
resolves #15.