Made require_relative without tailing \n match.
There was previously an issue with the regex I was using to match
require_relative calls and it was missing require_relative calls that didn't
have a \n or whitespace after the require_relative. Therefore, I changed the
regex to allow zero or more whitespace after the require_relative call. This
resolves #20.