Make default rake task run test suit.
I did this so that the project would be following the Travis CI and BoxCI
standards. Also, it is a good convention so that people don't have to remember
exactly how to run the suite for each project. Instead they just need to
remember to run the default rake task.